Fix build errors for native video support
- Fixed iOS NFloat to float conversion for VideoAspectRatio property
- Added Windows platform stub implementation for NativeAd with all required properties
- Windows implementation throws PlatformNotSupportedException as AdMob is not supported on Windows

+++ src/Plugin.AdMob/Platforms/iOS/Native/NativeAd.cs
@@ -36,7 +36,7 @@ internal partial class NativeAd : NativeAdLoaderDelegate
 
     public double VideoDuration => _ad?.MediaContent?.Duration ?? 0.0;
 
-    public float VideoAspectRatio => _ad?.MediaContent?.AspectRatio ?? 0.0f;
+    public float VideoAspectRatio => (float)(_ad?.MediaContent?.AspectRatio ?? 0.0f);
 
     public void Load()
     {
